Refactor, eliminating complexity
