+++ .gitignore
@@ -18,4 +18,6 @@ test/tmp
 test/version_tmp
 tmp
 matchers.komodoproject
+
 .idea/
+*.sublime*
