+++ .travis.yml
@@ -1,5 +1,4 @@
 language: ruby
 rvm:
-  - "2.0"
   - "2.1"
   - "2.2"
