+++ Rakefile
@@ -1,5 +1,4 @@
-require "bundler/gem_tasks"
-
+require 'bundler/gem_tasks'
 require 'rspec/core/rake_task'
 
 RSpec::Core::RakeTask.new('spec')
