+++ lib/rspec/webservice_matchers/version.rb
@@ -1,5 +1,5 @@
 module RSpec
   module WebserviceMatchers
-    VERSION = '4.0.0'
+    VERSION = '4.0.2'
   end
 end
