+++ rspec-webservice_matchers.gemspec
@@ -17,11 +17,12 @@ Gem::Specification.new do |spec|
   spec.executables   = spec.files.grep(/^bin\//) { |f| File.basename(f) }
   spec.test_files    = spec.files.grep(/^(test|spec|features)\//)
   spec.require_paths = ['lib']
+  spec.required_ruby_version = '>= 2.1.0'
 
   spec.add_development_dependency 'bundler', '~> 1.3'
   spec.add_development_dependency 'rake'
   spec.add_development_dependency 'pry'
-  spec.add_development_dependency 'webmock', '>= 1.20.4'
+  spec.add_development_dependency 'webmock'
 
   spec.add_runtime_dependency 'rspec', '~> 3.0'
   spec.add_runtime_dependency 'excon'
