+++ spec/rspec/webservice_matchers/public_api_spec.rb
@@ -1,13 +1,14 @@
 require 'spec_helper'
-require 'rspec/webservice_matchers'
+require 'rspec/webservice_matchers/util'
+include RSpec::WebserviceMatchers
 
 describe '#up?' do
   it 'follows redirects when necessary' do
-    expect(RSpec::WebserviceMatchers.up?('perm-redirector.com')).to be_truthy
-    expect(RSpec::WebserviceMatchers.up?('temp-redirector.org')).to be_truthy
+    expect(Util.up?('perm-redirector.com')).to be_truthy
+    expect(Util.up?('temp-redirector.org')).to be_truthy
   end
 
   it 'retries timeout errors once' do
-    expect(RSpec::WebserviceMatchers.up?('http://www.timeout-once.com')).to be_truthy
+    expect(Util.up?('http://www.timeout-once.com')).to be_truthy
   end
 end
